+++ src/middleware/validateSession.ts
@@ -7,10 +7,11 @@ const validateSession = async (req: Request, res: Response, next: NextFunction) 
   if (req.path === "/doc") {
     return next();
   }
-  if (req.path === "/users/login") {
+  if (req.path === "/login") {
     return next();
   }
   const sessionToken = req.headers.authorization;
+  process.stdout.write(sessionToken as string);
   if (!sessionToken) return res.status(401).send({ message: "No token provided." });
   return next();
 }
