+++ src/resources/users/user.service.ts
@@ -1,8 +1,8 @@
 /** @module UserService */
 import usersRepo from './user.memory.repository';
 import taskRepo from '../tasks/task.memory.repository';
-import { User } from "../../entity/User"
-import { UserDTO, JwtToken } from "../../common/types"
+import { User } from "../../entity/User";
+import { UserDTO } from "../../common/types";
 
 const getAll = async (): Promise<Array<User>> => usersRepo.getAll();
 
@@ -17,9 +17,4 @@ const deleteUser = async (id: string): Promise<void> => {
   await usersRepo.deleteUser(id);
 };
 
-const authorizeUser = async (login: string, password: string): Promise<JwtToken|undefined> => {
-  const token = await usersRepo.authorizeUser(login, password);
-  return token;
-}
-
-export default { getAll, getOne, createUser, updateUser, deleteUser, authorizeUser };
+export default { getAll, getOne, createUser, updateUser, deleteUser };
